1.0.0.130 - Aprimoramentos para melhor funcionalidade e utilização de LogSystem. - Pequenos ajustes em OutputByteArray, GenericOutput e BitWise8.

+++ Primitive/src/org/diverproject/util/BitWise8.java
@@ -68,7 +68,7 @@ public class BitWise8
 	 * @return aquisição do valor numérico inteiro das propriedades.
 	 */
 
-	public int getValue()
+	public byte getValue()
 	{
 		return value;
 	}
